PhotoSwipe can now group multiple media entities as one gallery.
This works by adding an attribute to entity reference fields that have
media entities as the target type, which photoswipe.field.js looks for
to build the gallery, which takes precedence over the image field
gallery grouping. In the future, a dedicated setting on the entity
reference field display may be added, but for the time being, this is
controlled via the media entity's image field settings.